- Sécurité & Rôles :     - Mise en place de la hiérarchie des rôles (ROLE_USER, ROLE_ADMIN).     - Création de Voters personnalisés pour la gestion des permissions (Suppression de produits).     - Sécurisation des routes et accès conditionnels dans les templates Twig.
- Interface d'Administration (EasyAdmin 5) :
    - Configuration du DashboardController avec accès restreint aux admins.
    - Création des CRUD Controllers pour les entités Produit et User.
    - Personnalisation du menu de navigation et redirection automatique.

- Design & UX :
    - Intégration de Bootstrap 5 avec un thème 'Dark Mode' cohérent entre le site et l'admin.
    - Refonte de la liste des produits en système de 'Cards' modernes.

- Mot de passe oublié :
    - Initialisation du bundle ResetPassword.
    - Configuration du composant Mailer pour l'envoi des emails de récupération.

+++ src/Controller/ProduitController.php
@@ -32,6 +32,9 @@ final class ProduitController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            // L'utilisateur connecté devient le propriétaire
+            $produit->setUser($this->getUser());
+            
             $em->persist($produit);
             $em->flush();
 
@@ -47,6 +50,9 @@ final class ProduitController extends AbstractController
     #[Route('/supprimer/{id}', name: 'produit_supprimer', methods: ['POST'])]
     public function supprimer(Produit $produit, EntityManagerInterface $em, Request $request): Response
     {
+        // MODIFICATION ICI : On demande au ProduitVoter si l'utilisateur a le droit
+        $this->denyAccessUnlessGranted('PRODUIT_DELETE', $produit);
+
         // On vérifie le jeton CSRF pour la sécurité
         if ($this->isCsrfTokenValid('delete'.$produit->getId(), $request->getPayload()->getString('_token'))) {
             $em->remove($produit);
